refactor: extract SandboxRepository
