+++ packages/control-plane/src/session/repository.test.ts
@@ -351,164 +351,4 @@ describe("SessionRepository", () => {
       expect(mock.calls).toHaveLength(3);
     });
   });
-
-  // === SANDBOX ===
-
-  describe("getSandbox", () => {
-    it("returns null when no sandbox exists", () => {
-      mock.setData(`SELECT * FROM sandbox LIMIT 1`, []);
-      expect(repo.getSandbox()).toBeNull();
-    });
-
-    it("returns sandbox when it exists", () => {
-      const sandbox = { id: "sb-1", status: "ready" };
-      mock.setData(`SELECT * FROM sandbox LIMIT 1`, [sandbox]);
-      expect(repo.getSandbox()).toEqual(sandbox);
-    });
-  });
-
-  describe("createSandbox", () => {
-    it("creates sandbox with correct parameters", () => {
-      repo.createSandbox({
-        id: "sb-1",
-        status: "pending",
-        gitSyncStatus: "pending",
-        createdAt: 1000,
-      });
-
-      expect(mock.calls.length).toBe(1);
-      expect(mock.calls[0].query).toContain("INSERT INTO sandbox");
-      expect(mock.calls[0].params).toEqual(["sb-1", "pending", "pending", 1000]);
-    });
-  });
-
-  describe("updateSandboxStatus", () => {
-    it("updates status", () => {
-      repo.updateSandboxStatus("ready");
-
-      expect(mock.calls.length).toBe(1);
-      expect(mock.calls[0].query).toContain("UPDATE sandbox SET status");
-      expect(mock.calls[0].params).toEqual(["ready"]);
-    });
-  });
-
-  describe("updateSandboxForSpawn", () => {
-    it("sets all spawn fields atomically", () => {
-      repo.updateSandboxForSpawn({
-        status: "spawning",
-        createdAt: 1000,
-        authTokenHash: "token-hash-123",
-        modalSandboxId: "modal-sb-1",
-      });
-
-      expect(mock.calls.length).toBe(1);
-      expect(mock.calls[0].query).toContain("UPDATE sandbox SET");
-      expect(mock.calls[0].query).toContain("status");
-      expect(mock.calls[0].query).toContain("auth_token_hash");
-      expect(mock.calls[0].query).toContain("modal_sandbox_id");
-      expect(mock.calls[0].query).toContain("auth_token = NULL");
-      expect(mock.calls[0].query).toContain("modal_object_id = NULL");
-      expect(mock.calls[0].query).toContain("vnc_url = NULL");
-      expect(mock.calls[0].query).toContain("vnc_password = NULL");
-      expect(mock.calls[0].params).toEqual(["spawning", 1000, "token-hash-123", "modal-sb-1"]);
-    });
-  });
-
-  describe("updateSandboxModalObjectId", () => {
-    it("updates modal object ID", () => {
-      repo.updateSandboxModalObjectId("obj-123");
-
-      expect(mock.calls.length).toBe(1);
-      expect(mock.calls[0].query).toContain("UPDATE sandbox SET modal_object_id");
-      expect(mock.calls[0].params).toEqual(["obj-123"]);
-    });
-  });
-
-  describe("updateSandboxSnapshotImageId", () => {
-    it("updates snapshot image ID for specific sandbox", () => {
-      repo.updateSandboxSnapshotImageId("sb-1", "img-123");
-
-      expect(mock.calls.length).toBe(1);
-      expect(mock.calls[0].query).toContain("UPDATE sandbox SET snapshot_image_id");
-      expect(mock.calls[0].params).toEqual(["img-123", "sb-1"]);
-    });
-  });
-
-  describe("updateSandboxHeartbeat", () => {
-    it("updates heartbeat timestamp", () => {
-      repo.updateSandboxHeartbeat(5000);
-
-      expect(mock.calls.length).toBe(1);
-      expect(mock.calls[0].query).toContain("UPDATE sandbox SET last_heartbeat");
-      expect(mock.calls[0].params).toEqual([5000]);
-    });
-  });
-
-  describe("updateSandboxLastActivity", () => {
-    it("updates activity timestamp", () => {
-      repo.updateSandboxLastActivity(6000);
-
-      expect(mock.calls.length).toBe(1);
-      expect(mock.calls[0].query).toContain("UPDATE sandbox SET last_activity");
-      expect(mock.calls[0].params).toEqual([6000]);
-    });
-  });
-
-  describe("updateSandboxGitSyncStatus", () => {
-    it("updates git sync status", () => {
-      repo.updateSandboxGitSyncStatus("completed");
-
-      expect(mock.calls.length).toBe(1);
-      expect(mock.calls[0].query).toContain("UPDATE sandbox SET git_sync_status");
-      expect(mock.calls[0].params).toEqual(["completed"]);
-    });
-  });
-
-  describe("updateSandboxSpawnError", () => {
-    it("updates spawn error fields", () => {
-      repo.updateSandboxSpawnError("Failed to spawn sandbox", 123456);
-
-      expect(mock.calls.length).toBe(1);
-      expect(mock.calls[0].query).toContain("UPDATE sandbox SET last_spawn_error");
-      expect(mock.calls[0].params).toEqual(["Failed to spawn sandbox", 123456]);
-    });
-  });
-
-  describe("VNC access", () => {
-    it("stores and clears VNC credentials", () => {
-      repo.updateSandboxVnc("https://vnc.test", "encrypted-password");
-      repo.clearSandboxVnc();
-
-      expect(mock.calls[0].query).toContain("SET vnc_url = ?, vnc_password = ?");
-      expect(mock.calls[0].params).toEqual(["https://vnc.test", "encrypted-password"]);
-      expect(mock.calls[1].query).toContain("SET vnc_url = NULL, vnc_password = NULL");
-    });
-
-    it("can clear only the VNC URL", () => {
-      repo.clearSandboxVncUrl();
-
-      expect(mock.calls[0].query).toContain("SET vnc_url = NULL");
-      expect(mock.calls[0].query).not.toContain("vnc_password");
-    });
-  });
-
-  describe("resetCircuitBreaker", () => {
-    it("resets failure count to zero", () => {
-      repo.resetCircuitBreaker();
-
-      expect(mock.calls.length).toBe(1);
-      expect(mock.calls[0].query).toContain("spawn_failure_count = 0");
-    });
-  });
-
-  describe("incrementCircuitBreakerFailure", () => {
-    it("increments count and sets timestamp", () => {
-      repo.incrementCircuitBreakerFailure(7000);
-
-      expect(mock.calls.length).toBe(1);
-      expect(mock.calls[0].query).toContain("spawn_failure_count = COALESCE");
-      expect(mock.calls[0].query).toContain("last_spawn_failure");
-      expect(mock.calls[0].params).toEqual([7000]);
-    });
-  });
 });
