+++ packages/control-plane/src/session/sandbox-events.test.ts
@@ -6,6 +6,7 @@ import type { ServerMessage } from "@open-inspect/shared/types/server-messages";
 import type { CallbackNotificationService } from "./callback-notification-service";
 import type { SessionDiffService } from "./diffs/service";
 import type { SessionRepository } from "./repository";
+import type { SandboxRepository } from "./sandbox-repository";
 import type { ArtifactRepository } from "./artifact-repository";
 import type { EventRepository } from "./event-repository";
 import type { MessageRepository } from "./message-repository";
@@ -86,6 +87,7 @@ function createProcessor() {
     { waitUntil } as unknown as DurableObjectState,
     () => log,
     repository as unknown as SessionRepository,
+    repository as unknown as SandboxRepository,
     repository as unknown as MessageRepository,
     eventRepository,
     artifactRepository,
